iOS Support, rename macOS+iOS module to "darwin"
