+++ battery/Cargo.toml
@@ -25,7 +25,7 @@ uom = { version = "~0.28", features = ["autoconvert", "f32", "si"] }
 [target.'cfg(target_os = "linux")'.dependencies]
 lazycell = "~1.2"
 
-[target.'cfg(target_os = "macos")'.dependencies]
+[target.'cfg(any(target_os = "macos", target_os = "ios"))'.dependencies]
 libc = "^0.2"
 mach = "^0.3"
 core-foundation = "~0.7"
