+++ battery/src/platform/darwin/device.rs
