+++ battery/src/platform/darwin/iokit/errors.rs
