+++ battery/src/platform/darwin/iokit/mod.rs
