+++ battery/src/platform/darwin/iokit/power_source.rs
