+++ battery/src/platform/darwin/iokit/sys.rs
