+++ battery/src/platform/darwin/iokit/wrappers.rs
