+++ battery/src/platform/darwin/iterator.rs
