+++ battery/src/platform/darwin/manager.rs
