+++ battery/src/platform/darwin/mod.rs
