+++ battery/src/platform/darwin/tests.rs
