+++ battery/src/platform/darwin/traits.rs
